% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_datasets.R
\docType{data}
\name{example_tree_data}
\alias{example_tree_data}
\title{Example datasets}
\format{
Tree data corresponding to three forest plots. Data frame with 40 rows (tree records) and columns:
\itemize{
  \item{\code{ID}: Forest plot code, defined as the combination of province code and "estadillo".}
  \item{\code{Species}: Tree species code.}
  \item{\code{N}: Tree density (individuals per hectare) that the record represents.}
  \item{\code{DBH}: Tree diameter at breast height (cm).}
  \item{\code{H}: Tree height (m)}
}
}
\description{
Data sets given to illustrate input data structures.
}
\references{
Project reference.
}
\author{
Miquel De Cáceres
}
\keyword{data}
