% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationDatasets.R
\docType{data}
\name{exampleTreeData}
\alias{exampleTreeData}
\title{Example datasets}
\description{
Data sets given to illustrate input data structures.
}
\details{
\itemize{
  \item{\code{exampleTreeData} Tree data corresponding to three forest plots. Data frame with 40 rows (tree records) and columns:
    \itemize{
        \item{\code{Provincia}: Province.}
        \item{\code{Estadillo}: Plot code within province.}
        \item{\code{ID}: Unique plot code within IFN.}
        \item{\code{Species}: Tree species code.}
        \item{\code{N}: Tree density (individuals per hectare) that the record represents.}
        \item{\code{DBH}: Tree diameter at breast height (cm).}
        \item{\code{H}: Tree height (m)}
    }
  }
}
}
\references{
Project reference.
}
\author{
Miquel De Cáceres
}
\keyword{data}
