% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNallometry-package.R
\docType{package}
\name{IFNallometry-package}
\alias{IFNallometry-package}
\alias{IFNallometry}
\title{IFNallometry: IFN volume and biomass allometry models}
\description{
Routines for calculating volume and biomass for Spanish Forest Inventory (IFN) plots.
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/IFNallometry/index.html}} }
}
\author{
Maintainer: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
[ORCID](https://orcid.org/0000-0001-7132-2080)
}
\keyword{internal}
