% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNbiomass.R
\name{IFNbiomass}
\alias{IFNbiomass}
\title{Biomass allometry models}
\usage{
IFNbiomass(x, as.CO2 = FALSE, area = NA, DBHclasses = NULL, verbose = FALSE)
}
\arguments{
\item{x}{A data frame with tree records in rows and columns 'Species', 'DBH' (in cm), 'H' (in m) and 'N' (ha-1)}

\item{as.CO2}{Flag to indicate output as kg of CO2 / ha instead of kg of dry weight / ha. Percentage of carbon per dry weight biomass by species are
taken from Montero et al. (2005) (in turn, from Ibáñez et al. 2002).}

\item{area}{Either 'Atlantic' or 'Mediterranean' to specify allometric equations specific to the area (for Pinus pinaster)}

\item{DBHclasses}{A numeric vector of DBH class limits (see breaks in function \code{cut}), used to group results by DBH class (in addition to species and plot). If \code{DBHclasses = NULL} then grouping on the basis of DBH classes is not performed.}

\item{verbose}{A flag to indicate console output of the biomass calculation process}
}
\value{
If \code{DBHclasses = NULL}, function \code{IFNbiomass} returns a data frame with as many rows as tree records in \code{x} and columns 'ID', 'Species', 'Roots', 'Stem', 'Branches', 'Leaves', 'Needles', 'Bark', 'Aerial' and 'Total'.
If \code{DBHclasses != NULL} then an extra column \code{DBHclass} is given and the data frame has less rows than tree records in \code{x}.
Function \code{IFNproducts} returns a data frame with the biomass of products (as well as that of stumps and slash), assuming trees have been felled down.
}
\description{
Static models for the biomass (kg/ha) of stem, roots, branches, leaves/needles and bark of species in the IFN
}
\details{
Function \code{IFNbiomass} determines the allometric equation for each biomass fraction (i.e. compartment) using species code.
All equations are calculated form height and diameter at breast height (DBH) of felled trees. Biomass values from equations refer to
individual trees, but the function multiplies the result by the density of individuals 'N', so the resulting value is in units of kg/ha.
Biomass of branches is the result of adding the result of equations calibrated for different branch diameters.
}
\examples{
data(exampleTreeData)

IFNbiomass(exampleTreeData)

# Groups the result by DBH clases
IFNbiomass(exampleTreeData,
           DBHclasses = seq(0, 120, by=5))

}
\references{
Diéguez-Aranda, U., A. Rojo Alboreca, F. Castedo-Dorado, J. G. Álvarez González, M. Barrio-Anta, F. Crecente-Campo, J. M. González González, C. Pérez-Cruzado, R. Rodríguez Soalleiro, C. A. López-Sánchez, M. Á. Balboa-Murias, J. J. Gorgoso Varela, and F. Sánchez Rodríguez (2009) Herramientas selvícolas para la gestión forestal sostenible en Galicia. Dirección Xeral de Montes, Consellería do Medio Rural, Xunta de Galicia.

Ibáñez JJ, Vayreda J., Gracia C. (2002) Metodología complementaria al Inventario Forestal Nacional en Catalunya. En: Bravo F.; del Río M.; del Peso C.
(eds.) El inventario Forestal Nacional. Elemento clave para la gestión forestal sostenible: 67-77. Fundación General de la Universidad de Valladolid
Montero G, Ruiz-Peinado R, Muñoz M (2005) Producción de biomasa y fijación de CO2 por los bosques españoles. Monografias INIA: Serie Forestal nº 13.

Ruiz-Peinado, R., M. Rio, and G. Montero (2011) New models for estimating the carbon sink capacity of Spanish. Forest Systems 20:176–188.

Ruiz-Peinado, R., G. Montero, and M. Del Rio (2012) Biomass models to estimate carbon stocks for hardwood tree species. Forest Systems 21:42.
}
