% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNbiomass.R
\name{IFNbiomass_medfate}
\alias{IFNbiomass_medfate}
\title{Wrapper biomass function for packages medfate and medfateland}
\usage{
IFNbiomass_medfate(x, area = NA, fraction = "aerial")
}
\arguments{
\item{x}{Data frame of tree data or \code{forest} object from package medfate.}

\item{area}{Either 'Atlantic' or 'Mediterranean' to specify allometric equations specific to the area (for Pinus pinaster)}

\item{fraction}{A string, either "total" or "aerial".}
}
\value{
A vector of biomass of each tree cohort (in Mg/ha of dry weight) to be used in medfate or medfateland packages.
}
\description{
Wrapper function to be used to calculate tree biomass for packages medfate and medfateland.
}
\details{
Values for parameter \code{x} will be supplied by the calling function (e.g. \code{modify_forest_structure}). Values for parameters
\code{area}, and \code{fraction} (if they defaults need to be changed) should be supplied using as a list to parameter \code{biomass_arguments}.
}
\seealso{
\code{\link{IFNbiomass}}
}
