% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNvolume.R
\name{IFNvolume}
\alias{IFNvolume}
\title{Volume allometry models}
\usage{
IFNvolume(
  x,
  IFN = c(3, 2),
  FC = 1:6,
  code_missing = "99",
  provinceFromID = FALSE,
  DBHclasses = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A data frame with tree records in rows and columns 'Provincia' (numeric), 'Species', 'DBH' (in cm), 'H' (in m) and 'N' (ha-1).
Additionally, the data frame may include a column 'FC' that specifies the cubic content form for each tree. In this case, the value of parameter 'FC' is overriden.}

\item{IFN}{Integer or integer pair to indicate order of preferred forest inventory:
(\code{IFN=2} indicates using IFN-2 only, \code{IFN=3} indicates using IFN-3 only, and \code{IFN=c(3,2)} indicates using IFN3 and if not available IFN2)}

\item{FC}{Vector of integers to indicate preferred cubic content forms (1 to 6)
\itemize{
   \item{Forma 1. Arboles fusiformes prácticamente en todo su fuste, con troncos maderables,
   limpios y derechos de más de 6 m, flecha inferior al 1% de su longitud, veta no torcida
   y diámetro normal mayor de 20 cm.}
   \item{Forma 2. Arboles que cumplan las cuatro condiciones siguientes: ser fusiformes, tener
   troncos maderables de 4 o más metros, ramificarse por la parte superior y no pertenecer
   a la forma 1.}
   \item{Forma 3. Arboles fusiformes pequeños, en los que el diámetro del fuste de 75 mm
   queda por debajo de los 4 m de altura.}
   \item{Forma 4. Árboles cuyo tronco principal se ramifica antes de los 4 m de altura y que
   pertenezcan a algunas de las siguientes especies 07, 12, 16, 23, 41, 42, 43, 44, 45, 46,
   47, 48, 49, 55, 56, 57, 66, 67, 71, 72, 74, 75, 79 y 94.}
   \item{Forma 5. Arboles cuyo tronco principal es tortuoso, está dañado o es muy ramoso, por
   lo que no admite la clasificación en formas 1, 2 o 3. También pies de altura de fuste
   menor de 4 m si son de especies diferentes a las de los códigos 4 y 6.}
   \item{Forma 6. Árboles descabezados o trasmochos a los que se les ha cortado la parte
   superior del tronco y las ramas en puntos próximos a su inserción en el tronco y que
   pertenezcan a algunas de las siguientes especies: 41, 42, 43, 55, 56, 71, 72 y 94.}
}}

\item{code_missing}{Species code to use when equations are not available for the species recorded ("99" is Otras frondosas)}

\item{provinceFromID}{A flag to indicate that province should be extracted from 'ID' assuming they are IFN plot codes.}

\item{DBHclasses}{A numeric vector of DBH class limits (see breaks in function \code{cut}), used to group results by DBH class (in addition to species and plot). If \code{DBHclasses = NULL} then grouping on the basis of DBH classes is not performed.}

\item{verbose}{A flag to indicate console output of the volume calculation process}
}
\value{
If \code{DBHclasses = NULL}, a data frame with as many rows as tree records in \code{x} and columns:
\itemize{
  \item{\code{ID}: Plot identifier}
  \item{\code{Species}: Species code}
  \item{\code{Name}: Species name}
  \item{\code{FC}: Cubic content form used to calculate volume.}
  \item{\code{VCC}: Volumen con corteza (m3/ha).}
  \item{\code{VSC}: Volumen sin corteza (m3/ha).}
  \item{\code{VLE}: Volumen de leñas (m3/ha).}
  \item{\code{IAVC}: Incremento anual del volumen con corteza (m3/ha/yr).}
}
If \code{DBHclasses != NULL} then an extra column \code{DBHclass} is given and the data frame has less rows than tree records in \code{x}.
}
\description{
Static models for the wood volume (VCC, VSC, VLE & IAVC) by species and province in the IFN
}
\details{
The volumetric equation used for each tree record depends on province, species, cubic content form and volume parameter (VCC, VSC, VLE and IAVC).
Volumes are given as per hectare (i.e. the result of the volumetric equation is multiplied by the density 'N'). If cubic content form is not given in 'x', then
the function iterates over the values of 'FC' until an equation is available.
}
\examples{
data(exampleTreeData)

IFNvolume(exampleTreeData)

# Groups the result by DBH clases
IFNvolume(exampleTreeData,
          DBHclasses = seq(0, 120, by=5))
}
