% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNvolume.R
\name{IFNvolume_forestindicators}
\alias{IFNvolume_forestindicators}
\title{Wrapper tree volume function for package forestindicators}
\usage{
IFNvolume_forestindicators(x, province, ...)
}
\arguments{
\item{x}{A data frame corresponding to plant_dynamic_input in package forestindicators}

\item{province}{A numeric code with Spanish province}

\item{...}{Parameters not used}
}
\value{
A vector of timber volumes per tree cohort (in m3/ha).
}
\description{
Wrapper tree volume function for package forestindicators
}
