% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNvolume.R
\name{IFNvolume_medfate}
\alias{IFNvolume_medfate}
\title{Wrapper volume function for packages medfate and medfateland}
\usage{
IFNvolume_medfate(x, SpParams, province, min_dbh = 7.5)
}
\arguments{
\item{x}{Data frame of tree data or \code{forest} object from package medfate.}

\item{SpParams}{Data frame of species parameters suitable for medfate package (not used).}

\item{province}{Spanish province code.}

\item{min_dbh}{Minimum diameter at breast height (in cm) for timber estimation.}
}
\value{
A vector of timber volumes per tree cohort (in m3/ha) to be used in medfateland simulations.
}
\description{
Wrapper function to be used to calculate timber volumes with packages medfate and medfateland.
}
\details{
Values for parameters \code{x} and \code{SpParams} will be supplied by the simulation function (e.g. \code{fordyn_scenario}) during calculations. Values for parameters
\code{province} and \code{min_dbh} should be supplied using as a list to parameter \code{volume_arguments}. See documentation of package medfateland.
}
\seealso{
\code{\link{IFNvolume}}
}
