% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_datasets.R
\docType{data}
\name{provincias}
\alias{provincias}
\title{Spanish provinces}
\format{
A data frame with province names and their inclusion in autonomous communities
\itemize{
  \item{\code{Provincia}: Province name.}
  \item{\code{CCAA}: Autonomous community where province belongs.}
}
}
\description{
Data set of Spanish provinces.
}
\references{
Project reference.
}
\author{
Miquel De Cáceres
}
\keyword{data}
