% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_datasets.R
\docType{data}
\name{species_ifn}
\alias{species_ifn}
\title{IFN species}
\format{
A data frame with species names and their code in IFN
\itemize{
  \item{\code{IFNcode}: Numeric code used in IFN.}
  \item{\code{Species}: Species name.}
}
}
\description{
Data set of recognized species names
}
\references{
Project reference.
}
\author{
Miquel De Cáceres
}
\keyword{data}
