% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNbiomass.R
\name{IFNbiomass_forestindicators}
\alias{IFNbiomass_forestindicators}
\title{Wrapper biomass function for package forestindicators}
\usage{
IFNbiomass_forestindicators(x, area = NA, fraction = "total", ...)
}
\arguments{
\item{x}{A data frame corresponding to plant_dynamic_input in package forestindicators}

\item{area}{Either 'Atlantic' or 'Mediterranean' to specify allometric equations specific to the area (for Pinus pinaster)}

\item{fraction}{A string, either "total" (for total biomass), "stem" (for stem biomass), "branches" (for branch biomass), "aboveground" (for aboveground biomass) or "belowground" (for belowground biomass).}

\item{...}{Parameters not used}
}
\value{
A vector of biomass of each tree cohort (in Mg/ha of dry weight)
}
\description{
Wrapper biomass function for package forestindicators
}
